services/ingestion-service/requirements.txt
pyspark==4.1.1
python-dotenv==1.0.0
confluent-kafka==2.3.0
